Move deprecation class to util
