+++ griptape/drivers/__init__.py
@@ -1,7 +1,5 @@
-from types import ModuleType
-import warnings
+from griptape.utils.deprecation import DeprecationModuleWrapper
 import sys
-from typing import Any
 
 from .prompt import BasePromptDriver
 from .prompt.openai import OpenAiChatPromptDriver
@@ -236,36 +234,9 @@ __all__ = [
 ]
 
 
-class _DeprecationWarningModuleWrapper(ModuleType):
-    """Module wrapper that issues a deprecation warning when importing."""
-
-    __ignore_attrs__ = {
-        "__file__",
-        "__package__",
-        "__path__",
-        "__doc__",
-        "__all__",
-        "__name__",
-        "__loader__",
-        "__spec__",
-    }
-
-    def __init__(self, real_module: Any) -> None:
-        self._real_module = real_module
-
-    def __getattr__(self, name: str) -> Any:
-        if name in self.__ignore_attrs__:
-            return getattr(self._real_module, name)
-
-        warnings.warn(
-            "Importing from `griptape.drivers` is deprecated and will be removed in a future release. "
-            "Please import from the provider-specific package instead.\n"
-            "e.g., `from griptape.drivers import OpenAiChatPromptDriver` -> `from griptape.drivers.prompt.openai import OpenAiChatPromptDriver`",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
-        return getattr(self._real_module, name)
-
-
-sys.modules[__name__] = _DeprecationWarningModuleWrapper(sys.modules[__name__])
+sys.modules[__name__] = DeprecationModuleWrapper(
+    sys.modules[__name__],
+    deprecation_message="Importing from `griptape.drivers` is deprecated and will be removed in a future release. "
+    "Please import from the provider-specific package instead.\n"
+    "e.g., `from griptape.drivers import OpenAiChatPromptDriver` -> `from griptape.drivers.prompt.openai import OpenAiChatPromptDriver`",
+)
